Update annotation suggestions
